+++ core/build.gradle
@@ -1,0 +1,21 @@
+plugins {
+    id 'org.springframework.boot'
+
+}
+
+group = 'com.voyagerss'
+version = '0.1.0'
+
+repositories {
+    mavenCentral()
+}
+
+bootJar{
+    enabled=false
+}
+jar{
+    enabled=true
+}
+
+dependencies {
+}
