+++ core/src/main/resources/logback-spring.xml
@@ -1,0 +1,43 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<configuration>
+  <include resource="org/springframework/boot/logging/logback/base.xml"/>
+
+  <springProperty scope="context" name="APPLICATION_NAME" source="application.name"/>
+  <property name="HOME_DIR" value="${user.home}/logs/${APPLICATION_NAME}"/>
+  <property name="FILE_NAME" value="${HOME_DIR}/log"/>
+
+  <!--    Log File Location -->
+  <!-- FILE Appender -->
+  <appender name="FILE_LOG" class="ch.qos.logback.core.rolling.RollingFileAppender">
+    <file>${FILE_NAME}</file>
+    <rollingPolicy class="ch.qos.logback.core.rolling.TimeBasedRollingPolicy">
+      <fileNamePattern>${FILE_NAME}.%d{yyyyMMdd}.log</fileNamePattern>
+      <maxHistory>60</maxHistory>
+    </rollingPolicy>
+    <encoder>
+      <pattern>${FILE_LOG_PATTERN}</pattern>
+    </encoder>
+  </appender>
+
+  <!-- CONSOLE Appender -->
+  <!--    <appender name="CONSOLE_LOG" class="ch.qos.logback.core.ConsoleAppender">-->
+  <!--        <layout class="ch.qos.logback.classic.PatternLayout">-->
+  <!--            <pattern>${CONSOLE_LOG_PATTERN}</pattern>-->
+  <!--        </layout>-->
+  <!--    </appender>-->
+
+  <!--    Profile Log Level Setting -->
+  <!-- TRACE > DEBUG > INFO > WARN > ERROR -->
+  <springProfile name="local">
+    <logger name="sun.rmi" level="INFO"/>
+    <logger name="javax.management" level="INFO"/>
+    <logger name="org.mongodb.driver" level="INFO"/>
+    <root level="DEBUG"/>
+  </springProfile>
+  <springProfile name="beta | prod">
+    <root level="INFO">
+      <appender-ref ref="FILE_LOG"/>
+    </root>
+  </springProfile>
+
+</configuration>
