+++ service/src/main/java/com/financial/service/repository/KrBankDataRepository.java
@@ -1,0 +1,11 @@
+package com.voyagerss.financial.repository;
+
+import com.voyagerss.financial.domain.KrBankData;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public
+interface KrBankDataRepository extends JpaRepository<KrBankData, Long>{
+
+}
