+++ service/src/main/resources/application.yml
@@ -1,0 +1,14 @@
+springdoc:
+    version: '1.0.0'
+    api-docs:
+        groups:
+            enabled: true
+    default-consumes-media-type: application/json
+    default-produces-media-type: application/json
+    swagger-ui:
+        operations-sorter: alpha
+        tags-sorter: alpha
+        path: /swagger-ui.html
+        disable-swagger-default-url: true
+        display-query-params-without-oauth2: true
+    paths-to-match: /**
