+++ service/src/test/java/com/financial/service/ServiceApplicationTests.java
@@ -1,0 +1,13 @@
+package com.financial.service;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class ServiceApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
