--- src/main/java/com/voyagerss/config/SchedulerConfig.java
@@ -1,22 +1,0 @@
-package com.voyagerss.config;
-
-import org.springframework.context.annotation.Configuration;
-import org.springframework.scheduling.annotation.SchedulingConfigurer;
-import org.springframework.scheduling.concurrent.ThreadPoolTaskScheduler;
-import org.springframework.scheduling.config.ScheduledTaskRegistrar;
-
-@Configuration
-public class SchedulerConfig implements SchedulingConfigurer {
-    private final int POOL_SIZE = 10;
-
-    @Override
-    public void configureTasks(ScheduledTaskRegistrar scheduledTaskRegistrar) {
-        ThreadPoolTaskScheduler threadPoolTaskScheduler = new ThreadPoolTaskScheduler();
-
-        threadPoolTaskScheduler.setPoolSize(POOL_SIZE);
-        threadPoolTaskScheduler.setThreadNamePrefix("my-scheduled-task-pool-");
-        threadPoolTaskScheduler.initialize();
-
-        scheduledTaskRegistrar.setTaskScheduler(threadPoolTaskScheduler);
-    }
-}
