--- src/main/resources/application.properties
@@ -1,41 +1,0 @@
-server.port=35000
-spring.application.name=charger
-spring.profiles=beta
-
-#devtools live reload
-spring.devtools.livereload.enabled=true
-
-
-#  DBCP
-#spring.datasource.initialization-mode=always
-#spring.datasource.type=org.apache.commons.dbcp.BasicDataSource
-spring.datasource.dbcp2.initial-size=7
-spring.datasource.dbcp2.max-total=20
-spring.datasource.dbcp2.pool-prepared-statements=true 
-
-spring.jpa.open-in-view=true
-#spring.jpa.hibernate.ddl-auto=update
-spring.jpa.hibernate.naming.physical-strategy=org.hibernate.boot.model.naming.PhysicalNamingStrategyStandardImpl
-#spring.jpa.properties.hibernate.id.new_generator_mappings=false
-
-#log
-logging.level.com.message=WARN
-logging.level.org.hibernate.SQL=debug
-logging.level.org.hibernate.type.descriptor.sql.BasicBinder=INFO
-
-#logging.path=C:/logs/charger/
-#logging.file=C:/logs/charger/charger.log
-logging.pattern.file=%d %p %c{1.} [%t] %m%n
-logging.pattern.console=%d{HH:mm:ss.SSS} [%t] %-5level %logger{36} - %msg%n
-
-spring.output.ansi.enabled=always
-
-## Multipart properties   # Enable multipart uploads
-spring.servlet.multipart.enabled=true 
-spring.servlet.multipart.maxFileSize= 200MB
-spring.servlet.multipart.maxrequestsize=215MB
-
-account.config=account.properties
-
-
-
