--- src/test/java/com/voyagerss/ServerApplicationTests.java
@@ -1,13 +1,0 @@
-package com.voyagerss;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class ServerApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
